pyproject.toml
[project]
name = "planogram"
version = "0.1.1"
description = ""
authors = [
    { name = "kyleryxn" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi >=0.116.2",
    "uvicorn >=0.35.0",
    "google-api-python-client >=2.182.0",
    "google-cloud-vision >=3.10.2",
    "google-auth-httplib2 >=0.2.0",
    "google-auth-oauthlib >=1.2.2",
    "jinja2 >=3.1.6",
    "psycopg[binary] >=3.2.10", # Postgres python drivers
    "redis >=6.4.0",
    "sqlalchemy[asyncio] >=2.0.43",
    "alembic >=1.16.5",
    "packaging >=25.0",
    "pydantic >=2.11.9",
    "pydantic-settings >=2.10.1",
]


[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
